# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TowTruckMove REST API",
    "description": ""
  },
  "host": "towtruckmove-api.onrender.com",
  "basePath": "/",
...
  "schemes": ["https"],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/driver/findOne/{driverId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findOneAssigned/{driverId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAll": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAllDriversByCompany/{companyId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAllDriversByDriverName/{driverName}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "driverName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAllDriversByVehicle/{vehicleId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAllDriversLicensed": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAllDriversByDate/{startDate}/{endDate}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/findAllAssignedDrivers": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/driver/create": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Email": {
                  "example": "any"
                },
                "Password\r": {
                  "example": "any"
                },
                "Password": {
                  "example": "any"
                },
                "CompanyId": {
                  "example": "any"
                },
                "DriverName": {
                  "example": "any"
                },
                "Phone": {
                  "example": "any"
                },
                "Address": {
                  "example": "any"
                },
                "DOB": {
                  "example": "any"
                },
                "City": {
                  "example": "any"
                },
                "Region": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "LicenseNo": {
                  "example": "any"
                },
                "Rating": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/driver/updateFile": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "RefId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/driver/updateDriverPic": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "RefId": {
                  "example": "any"
                },
                "FileType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver/AssignDriverToVehicle": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "DriverId": {
                  "example": "any"
                },
                "VehicleId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/sendDriverRegistrationLink": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Email": {
                  "example": "any"
                },
                "DriverName": {
                  "example": "any"
                },
                "CompanyId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/update/{driverId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "DriverName": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "Phone": {
                  "example": "any"
                },
                "Address": {
                  "example": "any"
                },
                "DOB": {
                  "example": "any"
                },
                "City": {
                  "example": "any"
                },
                "Region": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "Licensed": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/delete/{driverId}": {
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/driver/deleteAll": {
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}